Add files via upload
AJ wrote:
I'm preparing api-v3.yaml to include it inside the next release of TEDAPI. Since the status of TEDWEB is unclear, api-v4.yaml is not ready.

You can find the api-v3.yaml here: https://github.com/arnaudjardine/eu-tedapi/blob/master/api-v3.yaml (feel free to clone it and create a PR on it if necessary).
I have deployed a version of this API into our demo environment to ease the review if needed: https://demo.tedapi.spikeseed.cloud/webjars/swagger-ui/index.html#/

It contains:
•	The new operation /v3/config/sdk-versions of CVS
•	The new operation /v3/notices/render-async of TEDVIEWER
•	The parameters transmittedAt and createdAt to sort the notice search
•	The grouping of operations
•	The dropdown v3/v4 on top right
•	Correction to link the NoticeResponse to the correct schema
•	noticeAuthorLang instead of noticeAuthorLocale
Be aware that the following information is removed in this API compare to the one in CVS/TEDVIEW
parameters:
- name: X-API-Key
  in: header
  description: TED API key identifying the author of the request
  required: true
  schema:
    type: string

The reason is quite simple, this parameter is hidden for the API users since TEDAPI only expect the Authorization header. This X-API-Key is added by the gateway based on the Authorization header.